test: fields
