+++ test.js
@@ -32,292 +32,6 @@ ${message || defaultMessage}
   }
 }
 
-const array = ['foo', 'bar']
-const string = JSON.stringify(array)
-assert.strictEqual(
-  sortPackageJson(array),
-  array,
-  'should not sort object that is not plain object',
-)
-assert.strictEqual(
-  sortPackageJson(string),
-  string,
-  'should not sort object that is not plain object',
-)
-
-// fields sort as object
-for (const field of [
-  'exports',
-  'bin',
-  'contributes',
-  'commitlint',
-  'lint-staged',
-  'config',
-  'nodemonConfig',
-  'browserify',
-  'babel',
-  'xo',
-  'ava',
-  'jest',
-  'mocha',
-  'nyc',
-  'engines',
-  'engineStrict',
-  'preferGlobal',
-  'publishConfig',
-  'galleryBanner',
-]) {
-  testField(field, [
-    {
-      value: ['foo', 'bar'],
-      expect: ['foo', 'bar'],
-      message: `Should ignore \`${field}\` field that is not plain object.`,
-    },
-    {
-      value: {
-        foo: 'foo',
-        bar: 'bar',
-      },
-      expect: ['bar', 'foo'],
-      message: `Should sort \`${field}\` field.`,
-    },
-  ])
-}
-
-// simple disabled subKey sorting test
-for (const field of [
-  'name',
-  'displayName',
-  'version',
-  'description',
-  'sideEffects',
-  'files',
-  'categories',
-  'keywords',
-  'qna',
-  'publisher',
-  'type',
-  'main',
-  'umd:main',
-  'jsdelivr',
-  'unpkg',
-  'module',
-  'source',
-  'jsnext:main',
-  'browser',
-  'types',
-  'typings',
-  'style',
-  'example',
-  'examplestyle',
-  'assets',
-  'man',
-  'workspaces',
-  'activationEvents',
-  'pre-commit',
-  'browserslist',
-  'stylelint',
-  'flat',
-  'os',
-  'cpu',
-  'icon',
-  'preview',
-  'markdown',
-]) {
-  testField(field, [
-    {
-      value: ['foo', 'bar'],
-      expect: ['foo', 'bar'],
-      message: `Should not sort \`${field}\` field.`,
-    },
-    {
-      value: {
-        foo: 'foo',
-        bar: 'bar',
-      },
-      expect: ['foo', 'bar'],
-      message: `Should not sort \`${field}\` field.`,
-    },
-  ])
-}
-
-testField('husky', [
-  {
-    property: 'hooks',
-    value: {
-      hooks: {
-        'commit-msg': '',
-        [UNKNOWN]: UNKNOWN,
-        'pre-commit': '',
-      },
-    },
-    expect: ['pre-commit', 'commit-msg', UNKNOWN],
-  },
-])
-
-testField('binary', [
-  {
-    value: {
-      [UNKNOWN]: UNKNOWN,
-      module_name: 'node_addon_example',
-      module_path:
-        './lib/binding/{configuration}/{node_abi}-{platform}-{arch}/',
-      remote_path: './{module_name}/v{version}/{configuration}/',
-      package_name:
-        '{module_name}-v{version}-{node_abi}-{platform}-{arch}.tar.gz',
-      host: 'https://node-pre-gyp-tests.s3-us-west-1.amazonaws.com',
-    },
-    expect: [
-      'module_name',
-      'module_path',
-      'remote_path',
-      'package_name',
-      'host',
-      UNKNOWN,
-    ],
-  },
-])
-
-testField('keywords', [
-  {
-    value: ['foo', 'foo'],
-    expect: ['foo'],
-    message: 'keywords should be unique.',
-  },
-])
-
-testField('files', [
-  {
-    value: ['foo', 'foo'],
-    expect: ['foo'],
-    message: 'files should be unique.',
-  },
-])
-
-testField('bugs', [
-  {
-    value: {
-      email: 'npm@keithcirkel.co.uk',
-      url: 'https://github.com/keithamus/sort-package-json/issues',
-    },
-    expect: ['url', 'email'],
-  },
-  {
-    value: {
-      [UNKNOWN]: UNKNOWN,
-      email: 'npm@keithcirkel.co.uk',
-      url: 'https://github.com/keithamus/sort-package-json/issues',
-    },
-    expect: ['url', 'email', UNKNOWN],
-  },
-])
-
-testField('repository', [
-  {
-    value: {
-      [UNKNOWN]: UNKNOWN,
-      url: 'https://github.com/keithamus/sort-package-json',
-      type: 'github',
-    },
-    expect: ['type', 'url', UNKNOWN],
-  },
-])
-
-testField('funding', [
-  {
-    value: {
-      [UNKNOWN]: UNKNOWN,
-      url: 'https://github.com/keithamus/sort-package-json',
-      type: 'github',
-    },
-    expect: ['type', 'url', UNKNOWN],
-  },
-])
-
-testField('license', [
-  {
-    value: {
-      [UNKNOWN]: UNKNOWN,
-      url: 'https://example.com',
-      type: 'MIT',
-    },
-    expect: ['type', 'url', UNKNOWN],
-  },
-])
-
-testField('author', [
-  {
-    value: {
-      [UNKNOWN]: UNKNOWN,
-      url: 'http://keithcirkel.co.uk/',
-      name: 'Keith Cirkel',
-      email: 'npm@keithcirkel.co.uk',
-    },
-    expect: ['name', 'email', 'url', UNKNOWN],
-  },
-])
-
-// contributors
-assert.deepStrictEqual(
-  Object.keys(
-    sortPackageJson({
-      contributors: [
-        {
-          [UNKNOWN]: UNKNOWN,
-          url: 'http://keithcirkel.co.uk/',
-          name: 'Keith Cirkel',
-          email: 'npm@keithcirkel.co.uk',
-        },
-      ],
-    }).contributors[0],
-  ),
-  ['name', 'email', 'url', UNKNOWN],
-)
-assert.deepStrictEqual(
-  sortPackageJson({
-    contributors: ['foo', 'bar', ['foo', 'bar']],
-  }).contributors,
-  ['foo', 'bar', ['foo', 'bar']],
-)
-
-// badges
-assert.deepStrictEqual(
-  Object.keys(
-    sortPackageJson({
-      badges: [
-        {
-          [UNKNOWN]: UNKNOWN,
-          href: 'https://travis-ci.com/keithamus/sort-package-json.svg',
-          url: 'https://travis-ci.com/keithamus/sort-package-json',
-          description: 'sort-package-json build status',
-        },
-      ],
-    }).badges[0],
-  ),
-  ['description', 'url', 'href', UNKNOWN],
-)
-assert.deepStrictEqual(
-  sortPackageJson({
-    badges: ['foo', 'bar', ['foo', 'bar']],
-  }).badges,
-  ['foo', 'bar', ['foo', 'bar']],
-)
-
-testField('directories', [
-  {
-    value: {
-      [UNKNOWN]: UNKNOWN,
-      example: 'example',
-      man: 'man',
-      test: 'test',
-      doc: 'doc',
-      bin: 'bin',
-      lib: 'lib',
-    },
-    expect: ['lib', 'bin', 'man', 'doc', 'example', 'test', UNKNOWN],
-  },
-])
-
 // `resolutions` and `dependencies`
 for (const field of [
   'resolutions',
