+++ tests/snapshots/fields.js.md
@@ -1,0 +1,55 @@
+# Snapshot report for `tests/fields.js`
+
+The actual snapshot is saved in `fields.js.snap`.
+
+Generated by [AVA](https://ava.li).
+
+## badges
+
+> Should sort `badges[]`
+
+    `{␊
+      "badges": [␊
+        {␊
+          "description": "sort-package-json build status",␊
+          "url": "https://travis-ci.com/keithamus/sort-package-json",␊
+          "href": "https://travis-ci.com/keithamus/sort-package-json.svg",␊
+          "_": "this should still be the first element",␊
+          "a": "a",␊
+          "z": "z"␊
+        },␊
+        {␊
+          "description": "sort-package-json build status",␊
+          "url": "https://travis-ci.com/keithamus/sort-package-json",␊
+          "href": "https://travis-ci.com/keithamus/sort-package-json.svg",␊
+          "_": "this should still be the second element",␊
+          "a": "a",␊
+          "z": "z"␊
+        }␊
+      ]␊
+    }`
+
+## contributors
+
+> Should sort `contributors[]` as `PeopleField`
+
+    `{␊
+      "contributors": [␊
+        {␊
+          "name": "Keith Cirkel",␊
+          "email": "npm@keithcirkel.co.uk",␊
+          "url": "http://keithcirkel.co.uk/",␊
+          "_": "this should still be the first element",␊
+          "a": "a",␊
+          "z": "z"␊
+        },␊
+        {␊
+          "name": "Keith Cirkel",␊
+          "email": "npm@keithcirkel.co.uk",␊
+          "url": "http://keithcirkel.co.uk/",␊
+          "_": "this should still be the second element",␊
+          "a": "a",␊
+          "z": "z"␊
+        }␊
+      ]␊
+    }`
